requirements.txt
langchain-community
faiss-cpu
numpy
pypdf
python-dotenv
python-telegram-bot
pandas